Restructure the location of transformation scripts that handle data transformation to standard format
