+++ lib/transforms/lib/transformer/csv.py
@@ -1,15 +1,13 @@
 """
-CSV Dataset Parser A
-v 1.0-alpha
+CSV Dataset Transformer
 
-Parse the long format of traits by line, where the trait contains location and
-year. It splits a single file into N files, where N is the number of location
-and year combinations.
-
+This is the original transformer that was meant to transform the long format of
+traits by line, where the trait contains location and year. It splits a single
+file into N files, where N is the number of location and year combinations.
 
 Expected input:
   N files, M traits, P lines
-  Line/Pedigree, LocationYear 4-tuple,  Trait 1, Trait 2, Trait 3, Trait 4...Trait M
+  Line/Pedigree, Trait 1_LOYR, Trait 2_LOYR, Trait 3_LOYR, ...Trait M_LOYR
   Line 1, 1, 2, 3, 4, 5...M
   Line 2, 1, 2, 3, 4, 5...M
   ...
@@ -141,7 +139,7 @@ def process(args, delimiter = ','):
       dfs[filename]['data'] = df.filter(regex = pattern).set_index(list(df)[0]).dropna(how = 'all')
       # Rename columns to omit location-year pairs
       dfs[filename]['data'].columns = [ helpers.trait_to_column(t) for t in dfs[filename]['data'].columns ]
-
+      
     # Return the resultant dataframes
     return dfs
